requirements.txt
# Core LiveKit dependencies
livekit-agents>=1.1.4
livekit-plugins-openai>=1.1.4
livekit-plugins-silero>=1.1.4
livekit-plugins-google>=1.1.4
livekit-plugins-noise-cancellation>=0.2.4

# AI and Language Processing
mem0ai>=0.1.112
duckduckgo-search>=8.0.4
langchain_community>=0.3.26
langchain>=0.3.26
langchain-core>=0.3.66
langchain-text-splitters>=0.3.8
openai>=1.0.0

# Database & ORM
aiosqlite>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Processing and Analytics
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Web and API
requests>=2.31.0
httpx>=0.28.1
aiohttp>=3.9.0

# Environment and Configuration
python-dotenv>=1.0.0

# Date and Time
python-dateutil>=2.8.0

# JSON and Data Handling
orjson>=3.9.0

# Logging and Monitoring
coloredlogs>=15.0.1
humanfriendly>=10.0.0
structlog>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Audio Processing
av>=10.0.0
sounddevice>=0.4.0

# Google Services (for speech and other integrations)
google-cloud-speech>=2.33.0
google-cloud-texttospeech>=2.27.0
google-auth>=2.40.3
google-api-core>=2.25.1
google-genai>=1.23.0

# Additional Utilities
click>=8.2.1
colorama>=0.4.6
backoff>=2.2.1
cachetools>=5.5.2
